refactor(primitives): keep escrow payment classification in envelope
Amp-Thread-ID: https://ampcode.com/threads/T-019dcf0f-b633-761e-9c31-6c65eb7a3859
