+++ crates/contracts/src/precompiles/tip20_channel_escrow.rs
@@ -3,20 +3,9 @@ pub use ITIP20ChannelEscrow::{
     ITIP20ChannelEscrowEvents as TIP20ChannelEscrowEvent,
 };
 use alloy_primitives::{Address, address};
-use alloy_sol_types::{SolCall, SolInterface, SolType};
 
 pub const TIP20_CHANNEL_ESCROW_ADDRESS: Address =
     address!("0x4D50500000000000000000000000000000000000");
-
-const SECP256K1_SIGNATURE_LENGTH: usize = 65;
-const P256_SIGNATURE_TYPE: u8 = 0x01;
-const P256_SIGNATURE_LENGTH: usize = 130;
-const WEBAUTHN_SIGNATURE_TYPE: u8 = 0x02;
-const MIN_WEBAUTHN_SIGNATURE_LENGTH: usize = 129;
-const MAX_WEBAUTHN_SIGNATURE_LENGTH: usize = 2049;
-const KEYCHAIN_SIGNATURE_TYPE_V1: u8 = 0x03;
-const KEYCHAIN_SIGNATURE_TYPE_V2: u8 = 0x04;
-const KEYCHAIN_SIGNER_LENGTH: usize = 20;
 
 crate::sol! {
     #[derive(Debug, PartialEq, Eq)]
@@ -182,62 +171,6 @@ crate::sol! {
         error CloseNotReady();
         error DepositOverflow();
         error TransferFailed();
-    }
-}
-
-impl ITIP20ChannelEscrow::ITIP20ChannelEscrowCalls {
-    /// Returns `true` if `input` matches a channel escrow payment-lane selector and its calldata
-    /// is well-formed. `settle` and `close` also require a valid Tempo signature encoding.
-    pub fn is_payment(input: &[u8]) -> bool {
-        fn is_static_call<C: SolCall>(input: &[u8]) -> bool {
-            input.first_chunk::<4>() == Some(&C::SELECTOR)
-                && input.len()
-                    == 4 + <C::Parameters<'_> as SolType>::ENCODED_SIZE.unwrap_or_default()
-        }
-
-        if is_static_call::<ITIP20ChannelEscrow::openCall>(input)
-            || is_static_call::<ITIP20ChannelEscrow::topUpCall>(input)
-            || is_static_call::<ITIP20ChannelEscrow::requestCloseCall>(input)
-            || is_static_call::<ITIP20ChannelEscrow::withdrawCall>(input)
-        {
-            return true;
-        }
-
-        match Self::abi_decode(input) {
-            Ok(Self::settle(call)) => is_valid_tempo_signature_encoding(&call.signature),
-            Ok(Self::close(call)) => is_valid_tempo_signature_encoding(&call.signature),
-            _ => false,
-        }
-    }
-}
-
-fn is_valid_tempo_signature_encoding(signature: &[u8]) -> bool {
-    if is_valid_primitive_signature_encoding(signature) {
-        return true;
-    }
-
-    let Some((&signature_type, rest)) = signature.split_first() else {
-        return false;
-    };
-    if !matches!(
-        signature_type,
-        KEYCHAIN_SIGNATURE_TYPE_V1 | KEYCHAIN_SIGNATURE_TYPE_V2
-    ) || rest.len() <= KEYCHAIN_SIGNER_LENGTH
-    {
-        return false;
-    }
-
-    is_valid_primitive_signature_encoding(&rest[KEYCHAIN_SIGNER_LENGTH..])
-}
-
-fn is_valid_primitive_signature_encoding(signature: &[u8]) -> bool {
-    match signature.len() {
-        SECP256K1_SIGNATURE_LENGTH => true,
-        P256_SIGNATURE_LENGTH => signature.first() == Some(&P256_SIGNATURE_TYPE),
-        MIN_WEBAUTHN_SIGNATURE_LENGTH..=MAX_WEBAUTHN_SIGNATURE_LENGTH => {
-            signature.first() == Some(&WEBAUTHN_SIGNATURE_TYPE)
-        }
-        _ => false,
     }
 }
